requirements.txt
aiofiles
aiohttp
apscheduler
asyncio
bard
beautifulsoup4
bing_image_downloader
bs4
cloudscraper
curl-cffi
deepai
dnspython
emojis
Faker
ffmpeg-python
future
g4f
geopy
gitpython
google-api-python-client
gpytranslate
gtts
hachoir
heroku3
httpx[http2]==0.27.2
img2pdf
instaloader
langdetect
lexica-api
motor
nekosbest
numpy
openai==0.28.0
opencv-python
pickledb
pillow==10.3.0
psutil
lexica-api
ntgcalls
py-tgcalls
pycountry
pycryptodome
pydub
pyfiglet
pykeyboard
pyshorteners
python-dotenv
python-telegram-bot
python-whois
pyttsx3
pytube
pytz
pyyaml
qrcode
requests
search_engine_parser
speedtest-cli
SpeechRecognition
spotipy
telegraph
tgcrypto
unidecode
wget
youtube-search
youtube-search-python
git+https://github.com/KurimuzonAkuma/pyrogram.git
git+https://github.com/yt-dlp/yt-dlp.git
